requirements.txt
protobuf==6.31.1
pydantic==2.11.7
pydantic_ai==0.3.2
python-dotenv==1.1.0
